patch14: refactor col_nnz to Vec<u32>, add MTX round-trip + Seurat tests
- Change ScxCsr::col_nnz return type from Vec<i64> to Vec<u32>
- Propagate u32 type through scx-format, scx-accel, pyscx
- Update all call sites: backed.rs, lazy_transform.rs, filtering.rs,
  preprocessing.rs, projected_agg.rs, col_aggs.rs
- Add comprehensive MTX round-trip integration tests (6 tests):
  v2/v3 metadata parity, all codecs, multi-shard, float, empty matrix
- Add Seurat v4/v5 R-binding regression tests (8 tests):
  to_seurat, from_seurat round-trip, query results, filtered queries

Verified: cargo test, clippy clean, 789 pytest passed (identical to baseline)

+++ pyscx/src/accel/preprocessing.rs
@@ -359,7 +359,7 @@ pub fn calculate_qc_metrics<'py>(
     };
 
     // Compute per-gene total_counts and n_cells_by_counts.
-    let (gene_total_counts, n_cells): (Vec<f64>, Vec<i64>) = if prefer_format == "csc" {
+    let (gene_total_counts, n_cells): (Vec<f64>, Vec<u32>) = if prefer_format == "csc" {
         // CSC dispatch: capability gate + projected_agg twins.
         compute_gene_axis_csc(&x)?
     } else if is_backed {
@@ -380,7 +380,7 @@ pub fn calculate_qc_metrics<'py>(
                 .col_nnz_masked(kept)
                 .map_err(|e| PyRuntimeError::new_err(e.to_string()))?
                 .iter()
-                .map(|&v| v as i64)
+                .map(|&v| v as u32)
                 .collect(),
             None => backed
                 .backed
@@ -400,7 +400,7 @@ pub fn calculate_qc_metrics<'py>(
                 .col_nnz_masked(kept)
                 .map_err(|e| PyRuntimeError::new_err(e.to_string()))?
                 .iter()
-                .map(|&v| v as i64)
+                .map(|&v| v as u32)
                 .collect()
         } else {
             lazy.backed
@@ -816,7 +816,7 @@ fn gpu_log1p_dispatch(
 /// transforms / row deletion vector active) is delegated to
 /// `as_column_source()`. Honors `col_projection` if set on the
 /// dataset.
-fn compute_gene_axis_csc(x: &Bound<'_, PyAny>) -> PyResult<(Vec<f64>, Vec<i64>)> {
+fn compute_gene_axis_csc(x: &Bound<'_, PyAny>) -> PyResult<(Vec<f64>, Vec<u32>)> {
     if let Ok(backed) = x.extract::<PyRef<ScxBackedSparseDataset>>() {
         let source = backed.as_column_source().ok_or_else(|| {
             PyRuntimeError::new_err(
